updated cycle times to use integers

+++ classes/appliance.cpp
@@ -16,14 +16,14 @@ Appliance::Appliance(
 
 // Setters
 void Appliance::setName(const std::string& name) { name_ = name; }
-void Appliance::setCycleTime(double minutes) { cycle_time_ = minutes; }
+void Appliance::setCycleTime(int minutes) { cycle_time_ = minutes; }
 void Appliance::setWaterUsagePerMinute(const std::vector<double>& liters) { water_usage_per_minute_ = liters; }
 void Appliance::setInterruptable(bool interruptable) { interruptable_ = interruptable; }
 void Appliance::setTakesGreywater(bool takes_greywater) { takes_greywater_ = takes_greywater; }
 
 // Getters
 std::string Appliance::name() const { return name_; }
-double Appliance::cycleTime() const { return cycle_time_; }
+int Appliance::cycleTime() const { return cycle_time_; }
 const std::vector<double>& Appliance::waterUsagePerMinute() const { return water_usage_per_minute_; }
 bool Appliance::interruptable() const { return interruptable_; }
 bool Appliance::takesGreywater() const { return takes_greywater_; }
@@ -31,7 +31,7 @@ bool Appliance::takesGreywater() const { return takes_greywater_; }
 
 GreywaterAppliance::GreywaterAppliance(
     const std::string& name,
-    double cycle_time,
+    int cycle_time,
     const std::vector<double>& water_usage_per_minute,
     bool interruptable,
     double water_output_per_cycle
