+++ classes/appliance.hpp
@@ -13,21 +13,21 @@ class Appliance {
 
         // Setters
         void setName(const std::string& name);
-        void setCycleTime(double minutes);
+        void setCycleTime(int minutes);
         void setWaterUsagePerMinute(const std::vector<double>& liters);
         void setInterruptable(bool interruptable);
         void setTakesGreywater(bool takes_greywater);
 
         // Getters
         std::string name() const;
-        double cycleTime() const;
+        int cycleTime() const;
         const std::vector<double>& waterUsagePerMinute() const;
         bool interruptable() const;
         bool takesGreywater() const;
 
     protected:
         std::string name_;
-        double cycle_time_; // minutes
+        int cycle_time_; // minutes
         std::vector<double> water_usage_per_minute_; // appliances usually have bursts of water usage
         bool interruptable_; // some appliances, like washing machine, should not be stopped and left to sit for extended periods of time
         bool takes_greywater_;
@@ -38,7 +38,7 @@ class GreywaterAppliance: public Appliance {
     public:
         GreywaterAppliance(
             const std::string& name,
-            double cycle_time,
+            int cycle_time,
             const std::vector<double>& water_usage_per_minute,
             bool interruptable,
             double water_output_per_cycle
